requirements.txt
SpeechRecognition
pipwin
gTTS
google-generativeai
python-dotenv
streamlit
pyttsx3

-e .